+++ database/migrations/2023_09_19_090036_create_license_windows_table.php
@@ -16,6 +16,8 @@ return new class extends Migration
             $table->foreignId('customer_id')->constrained('customers')->onDelete('cascade');
             $table->foreignId('operating_system_id')->constrained('operating_systems')->onDelete('cascade')->nullable();
             $table->string('key');
+            $table->string('file_path')->nullable();
+            $table->string('file_name')->nullable();
             $table->timestamps();
             $table->softDeletes();
         });