+++ resources/views/licenseaccess/create.blade.php
@@ -5,5 +5,10 @@
 
         <x-create.singlerow label="Key" name="key" />
 
+        <x-input.label value="Datei" class="mt-2" />
+        <x-input.file name="file" />
+
+        <x-create.singlerow label="Datei Name" name="file_name" />
+
     </x-create.main>
 </x-app-layout>