+++ resources/views/licensesoftware/create.blade.php
@@ -9,5 +9,10 @@
 
         <x-create.singlerow label="Passwort" name="password" />
 
+        <x-input.label value="Datei" class="mt-2" />
+        <x-input.file name="file" />
+
+        <x-create.singlerow label="Datei Name" name="file_name" />
+
     </x-create.main>
 </x-app-layout>