+++ resources/views/licensesoftware/edit.blade.php
@@ -10,6 +10,11 @@
 
         <x-create.singlerow label="Passwort" name="password" default="{{ $licensesoftware->password }}" />
 
+        <x-input.label value="Datei" class="mt-2" />
+        <x-input.file name="file" />
+
+        <x-create.singlerow label="Datei Name" name="file_name" />
+
     </x-create.main>
 
     @can('licensesoftware_delete')